+++ inst/tinytest/test-sird.R
@@ -42,11 +42,8 @@ good_transmission_rate <- 0.9
 good_recovery_rate <- 0.3
 good_death_rate <- .1
 
-bad_name <- NA
-bad_prevalence <- NA
-bad_transmission_rate <- NA
-bad_recovery_rate <- NA
-bad_death_rate <- NA
+bad_name <- 10
+bad_numeric_input <- "not a number"
 
 expected_error_msg_str <- "must be a string"
 expected_error_msg_double <- "must be a double"
@@ -61,7 +58,7 @@ expect_error(sird_0 <- ModelSIRD(
 
 expect_error(sird_0 <- ModelSIRD(
   name = good_name,
-  prevalence = bad_prevalence,
+  prevalence = bad_numeric_input,
   transmission_rate = good_transmission_rate,
   recovery_rate = good_recovery_rate,
   death_rate = good_death_rate
@@ -70,7 +67,7 @@ expect_error(sird_0 <- ModelSIRD(
 expect_error(sird_0 <- ModelSIRD(
   name = good_name,
   prevalence = good_prevalence,
-  transmission_rate = bad_transmission_rate,
+  transmission_rate = bad_numeric_input,
   recovery_rate = good_recovery_rate,
   death_rate = good_death_rate
 ), expected_error_msg_double)
@@ -79,7 +76,7 @@ expect_error(sird_0 <- ModelSIRD(
   name = good_name,
   prevalence = good_prevalence,
   transmission_rate = good_transmission_rate,
-  recovery_rate = bad_recovery_rate,
+  recovery_rate = bad_numeric_input,
   death_rate = good_death_rate
 ), expected_error_msg_double)
 
@@ -88,5 +85,5 @@ expect_error(sird_0 <- ModelSIRD(
   prevalence = good_prevalence,
   transmission_rate = good_transmission_rate,
   recovery_rate = good_recovery_rate,
-  death_rate = bad_death_rate
+  death_rate = bad_numeric_input
 ), expected_error_msg_double)