+++ inst/tinytest/test-sirdconn.R
@@ -25,13 +25,8 @@ expect_inherits(sirdconn_0, "epiworld_model")
 expect_length(class(sirdconn_0), 2)
 
 # Check functions fail with invalid inputs -------------------------------------
-bad_name <- NA
-bad_n <- NA
-bad_prevalence <- NA
-bad_contact_rate <- NA
-bad_transmission_rate <- NA
-bad_recovery_rate <- NA
-bad_death_rate <- NA
+bad_name <- 10
+bad_numeric_input <- "not a number"
 
 expected_error_msg_str <- "must be a string"
 expected_error_msg_int <- "must be an integer"
@@ -49,7 +44,7 @@ expect_error(sirdconn_0 <- ModelSIRDCONN(
 
 expect_error(sirdconn_0 <- ModelSIRDCONN(
   name = good_name,
-  n = bad_n,
+  n = bad_numeric_input,
   prevalence = good_prevalence,
   contact_rate = good_contact_rate,
   transmission_rate = good_transmission_rate,
@@ -60,7 +55,7 @@ expect_error(sirdconn_0 <- ModelSIRDCONN(
 expect_error(sirdconn_0 <- ModelSIRDCONN(
   name = good_name,
   n = good_n,
-  prevalence = bad_prevalence,
+  prevalence = bad_numeric_input,
   contact_rate = good_contact_rate,
   transmission_rate = good_transmission_rate,
   recovery_rate = good_recovery_rate,
@@ -71,7 +66,7 @@ expect_error(sirdconn_0 <- ModelSIRDCONN(
   name = good_name,
   n = good_n,
   prevalence = good_prevalence,
-  contact_rate = bad_contact_rate,
+  contact_rate = bad_numeric_input,
   transmission_rate = good_transmission_rate,
   recovery_rate = good_recovery_rate,
   death_rate = good_death_rate
@@ -82,7 +77,7 @@ expect_error(sirdconn_0 <- ModelSIRDCONN(
   n = good_n,
   prevalence = good_prevalence,
   contact_rate = good_contact_rate,
-  transmission_rate = bad_transmission_rate,
+  transmission_rate = bad_numeric_input,
   recovery_rate = good_recovery_rate,
   death_rate = good_death_rate
 ), expected_error_msg_double)
@@ -93,7 +88,7 @@ expect_error(sirdconn_0 <- ModelSIRDCONN(
   prevalence = good_prevalence,
   contact_rate = good_contact_rate,
   transmission_rate = good_transmission_rate,
-  recovery_rate = bad_recovery_rate,
+  recovery_rate = bad_numeric_input,
   death_rate = good_death_rate
 ), expected_error_msg_double)
 
@@ -104,5 +99,5 @@ expect_error(sirdconn_0 <- ModelSIRDCONN(
   contact_rate = good_contact_rate,
   transmission_rate = good_transmission_rate,
   recovery_rate = good_recovery_rate,
-  death_rate = bad_death_rate
+  death_rate = bad_numeric_input
 ), expected_error_msg_double)