+++ inst/tinytest/test-sisd.R
@@ -21,11 +21,8 @@ expect_inherits(sisd_0, "epiworld_model")
 expect_length(class(sisd_0), 2)
 
 # Check functions fail with invalid inputs -------------------------------------
-bad_name                <- NA
-bad_prevalence          <- NA
-bad_transmission_rate   <- NA
-bad_recovery_rate       <- NA
-bad_death_rate          <- NA
+bad_name                <- 10
+bad_numeric_input       <- "not a number"
 
 expected_error_msg_str    <- "must be a string"
 expected_error_msg_double <- "must be a double"
@@ -40,7 +37,7 @@ expect_error(sisd_0 <- ModelSISD(
 
 expect_error(sisd_0 <- ModelSISD(
   name                = good_name,
-  prevalence          = bad_prevalence,
+  prevalence          = bad_numeric_input,
   transmission_rate   = good_transmission_rate,
   recovery_rate       = good_recovery_rate,
   death_rate          = good_death_rate
@@ -49,7 +46,7 @@ expect_error(sisd_0 <- ModelSISD(
 expect_error(sisd_0 <- ModelSISD(
   name                = good_name,
   prevalence          = good_prevalence,
-  transmission_rate   = bad_transmission_rate,
+  transmission_rate   = bad_numeric_input,
   recovery_rate       = good_recovery_rate,
   death_rate          = good_death_rate
 ), expected_error_msg_double)
@@ -58,7 +55,7 @@ expect_error(sisd_0 <- ModelSISD(
   name                = good_name,
   prevalence          = good_prevalence,
   transmission_rate   = good_transmission_rate,
-  recovery_rate       = bad_recovery_rate,
+  recovery_rate       = bad_numeric_input,
   death_rate          = good_death_rate
 ), expected_error_msg_double)
 
@@ -67,6 +64,6 @@ expect_error(sisd_0 <- ModelSISD(
   prevalence          = good_prevalence,
   transmission_rate   = good_transmission_rate,
   recovery_rate       = good_recovery_rate,
-  death_rate          = bad_death_rate
+  death_rate          = bad_numeric_input
 ), expected_error_msg_double)